[#28] Extract core modules from Otto class
- Extract 5 core modules: Router, FileSafety, Configuration, ErrorHandler, UriGenerator
- Use composition pattern with module inclusion for backward compatibility
- All 384 tests pass, maintaining complete API compatibility
- Reduces main Otto class from 767 lines to focused responsibilities



# Conflicts:
#	lib/otto.rb
